requirements.txt
asyncpg
structlog
pydantic
python-dotenv
fastapi
tortoise-orm
uvicorn